restructure progress service tests by methods
